fixed example code with ParseException and updated result
There was one ")" missing, producing an ParseException. I changed it to the same tree you used at the end of the article.
Furthermore, the example output did not match to the input. I changed the output to the actual output of the (corrected) input.
I guess in a former version of the article you experimented with a more complex tree, but changed it to a simpler one?

Anyway, thanks for the article!
For me, it has just the correct balance between not being too simple and not being too advanced. :)